Final-project_training/microservices/classification/requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3
python-docx==1.1.0
pandas==2.1.4
PyMuPDF==1.23.8
openpyxl==3.1.2
